Explore P0: move SourceKind/Confidence to enums/context_pack_enums
Relocate the two pack enums out of context_pack/types.go into enums/context_pack_enums/ (source_kind.go, confidence.go), matching the repo's enum-package convention (runner_enums-style: related named enums grouped in one *_enums package). types.go now imports context_pack_enums; ManifestEntry field types updated. Type identity unchanged, so kit/deployment-runner/deployment-server are unaffected — all rebuild clean.
